% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minka_obs_user.R
\name{get_minka_obs_user}
\alias{get_minka_obs_user}
\title{Download observations for a user}
\usage{
get_minka_obs_user(username, maxresults = 1000)
}
\arguments{
\item{username}{username of the Minka user to fetch records}

\item{maxresults}{the maximum number of results to return}
}
\value{
a list with full details on a given record
}
\description{
Get all the observations of a specific Minka user.
}
\examples{
\dontrun{
  m_obs <- get_minka_obs(query="Seabream")
  get_minka_obs_user(as.character(m_obs$user_login[1]))
}
}
